Split native review pane components
