+++ Sources/QuillCodeApp/QuillCodeReviewActionButton.swift
@@ -1,0 +1,21 @@
+import SwiftUI
+
+struct QuillCodeReviewActionButton: View {
+    var action: WorkspaceReviewActionSurface
+    var path: String
+    var onReviewAction: (WorkspaceReviewActionSurface) -> Void
+
+    var body: some View {
+        Button {
+            onReviewAction(action)
+        } label: {
+            Label(action.kind.title, systemImage: action.kind.systemImage)
+                .labelStyle(.iconOnly)
+                .frame(width: QuillCodeMetrics.minimumHitTarget, height: QuillCodeMetrics.minimumHitTarget)
+                .contentShape(Rectangle())
+        }
+        .buttonStyle(QuillCodePressableButtonStyle())
+        .help("\(action.kind.title) \(path)")
+        .foregroundStyle(action.kind == .restore || action.kind == .restoreHunk ? QuillCodePalette.yellow : QuillCodePalette.blue)
+    }
+}
